llrp: 🔄 Moved fill functions to properties
